🛠️ refactor: 타입 파일 이름 정리 (#54)
