+++ src/app/[locale]/(main)/page.tsx
@@ -26,30 +26,30 @@ export default function MainPage() {
 
   const influencerTest1 = {
     influencerId: '1',
-    name: '알간지',
-    imageSrc: '/assets/images/test/alganzi.png',
-    isVerified: false,
+    influencerName: '알간지',
+    influencerImageUrl: '/assets/images/test/alganzi.png',
+    isAuthenticated: false,
   };
 
   const influencerTest2 = {
     influencerId: '2',
-    name: '레오제이',
-    imageSrc: '/assets/images/test/leo.png',
-    isVerified: true,
+    influencerName: '레오제이',
+    influencerImageUrl: '/assets/images/test/leo.png',
+    isAuthenticated: true,
   };
 
   const influencerTest3 = {
     influencerId: '3',
-    name: '레오제이',
-    imageSrc: '/assets/images/test/alganzi.png',
-    isVerified: false,
+    influencerName: '레오제이',
+    influencerImageUrl: '/assets/images/test/alganzi.png',
+    isAuthenticated: false,
   };
 
   const influencerTest4 = {
     influencerId: '5',
-    name: '레오제이',
-    imageSrc: '/assets/images/test/leo.png',
-    isVerified: true,
+    influencerName: '레오제이',
+    influencerImageUrl: '/assets/images/test/leo.png',
+    isAuthenticated: true,
   };
   return (
     <div className="mt-6 w-full flex-col-center">
