+++ src/types/domain/boardType.ts
